Add a master-only Dockerfile
The new Dockerfile builds master as an archive! This
would be great for local builds for authors, as it
is only 377MB instead of 1.71GB which master + all
the archives are.

We could add this archive to docs-base and actually
use docs-base as what we actually serve as the live
docs.